+++ cli/runner/cmd/langs.go
@@ -8,7 +8,7 @@ import (
 	"fmt"
 
 	"github.com/runner-x/runner-x/cli/runner/client"
-	"github.com/runner-x/runner-x/engine/coderunner"
+	coderunner "github.com/runner-x/runner-x/engine/coderunner/v1"
 	"github.com/spf13/cobra"
 )
