+++ engine/coderunner/v1/runner_test.go
@@ -1,4 +1,4 @@
-package coderunner
+package v1
 
 import (
 	"errors"
